[Refactoring] Move ConvertToDoCatch to its own file
